+++ tests/invalid/global_variables/use_before_declaration.rs
@@ -2,8 +2,8 @@
 fail: 1
 */
 
-main(): int {
+main(): i32 {
     return foo;
 }
 
-foo: int = 3;
+foo: i64 = 3;
