+++ tests/invalid/loops/too_few_for_clauses.rs
@@ -2,8 +2,8 @@
 fail: 1
 */
 
-main(): int {
-    for (i: int = 2; i < 3)
+main(): i32 {
+    for (i: i64 = 2; i < 3)
         3;
     return 0;
 }
