+++ tests/invalid/variables/redefine.rs
@@ -2,8 +2,8 @@
 fail: 1
 */
 
-main(): int {
-    a: int = 1;
-    a: int = 2;
+main(): i32 {
+    a: i64 = 1;
+    a: i64 = 2;
     return a;
 }
