+++ tests/valid/boolean_operators/short_circuit_or.rs
@@ -3,9 +3,9 @@ return: 0
 skip: 1
 */
 
-main(): int {
-    a: int = 1;
-    b: int = 0;
+main(): i32 {
+    a: i64 = 1;
+    b: i64 = 0;
     a || (b = 5);
     return b;
 }
