+++ tests/valid/conditional_expression/ternary.rs
@@ -2,7 +2,7 @@
 return: 4
 */
 
-main(): int {
-    a: int = 0;
+main(): i32 {
+    a: i64 = 0;
     return a > -1 ? 4 : 5;
 }
