+++ tests/valid/conditional_statement/if_nested_4.rs
@@ -2,8 +2,8 @@
 return: 4
 */
 
-main(): int {
-    a: int = 0;
+main(): i32 {
+    a: i64 = 0;
     if (1)
         if (0)
             a = 3;
