+++ tests/valid/functions/fun_in_expr.rs
@@ -2,12 +2,12 @@
 return: 16
 */
 
-sum(a: int, b: int): int {
+sum(a: i64, b: i64): i64 {
     return a + b;
 }
 
-main(): int {
-    a: int = sum(1, 2) - (sum(1, 2) / 2) * 2;
-    b: int = 2*sum(3, 4) + sum(1, 2);
+main(): i32 {
+    a: i64 = sum(1, 2) - (sum(1, 2) / 2) * 2;
+    b: i64 = 2*sum(3, 4) + sum(1, 2);
     return b - a;
 }
