+++ tests/valid/loops/empty_expression.rs
@@ -2,10 +2,10 @@
 return: 3
 */
 
-main(): int {
-    i: int = 3;
+main(): i32 {
+    i: i64 = 3;
     ; 
-    for (i: int = 0; i < 10; i = i + 1)
+    for (i: i64 = 0; i < 10; i = i + 1)
         ;
     return i;
 }
