+++ tests/valid/loops/nested_while.rs
@@ -2,11 +2,11 @@
 return: 65
 */
 
-main(): int {
-    a: int = 1;
+main(): i32 {
+    a: i64 = 1;
 
     while (a / 3 < 20) {
-        b: int = 1;
+        b: i64 = 1;
         while (b < 10)
             b = b*2;
         a = a + b;
