+++ tests/valid/loops/while_single_statement.rs
@@ -2,8 +2,8 @@
 return: 6
 */
 
-main(): int {
-    a: int = 0;
+main(): i32 {
+    a: i64 = 0;
 
     while (a < 5)
         a = a + 2;
